+++ src/control_plane/shared_infrastructure/mongodb/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) use mongodb_logical_resource_plan::MongoDbLogicalResourcePlan;
+pub(crate) use mongodb_plan_error::MongoDbPlanError;
+pub(crate) use mongodb_shared_instance_plan::MongoDbSharedInstancePlan;
+pub(crate) use mongodb_shared_instance_plan_options::MongoDbSharedInstancePlanOptions;
+
+mod mongodb_logical_resource_plan;
+mod mongodb_plan_error;
+mod mongodb_shared_instance_plan;
+mod mongodb_shared_instance_plan_options;